recursion_sentry: Remove a cast

+++ asteria/recursion_sentry.cpp
@@ -8,7 +8,7 @@ namespace asteria {
 
 void
 Recursion_Sentry::
-do_throw_stack_overflow(size_t usage, size_t limit) const
+do_throw_stack_overflow(size_t usage, uint32_t limit) const
   {
     ASTERIA_THROW((
         "Stack overflow averted (stack usage `$1` exceeded `$2`)"),
